+++ data/scripts/constellations.py
@@ -5,12 +5,19 @@ import pandas as pd
 
 from shapely.geometry import Polygon
 
-from starplot.data import constellations
+from starplot.data.translations import language_name_column
 
+from constellation_hips import properties
 from data_settings import BUILD_PATH, RAW_PATH
+from translations import get_translations
 
 DATA_PATH = RAW_PATH / "iau"
 CRS = "+ellps=sphere +f=0 +proj=latlong +axis=wnu +a=6378137 +no_defs"
+
+translated = get_translations("constellation_names.csv")
+language_columns = [
+    (language, language_name_column(language)) for language in translated.keys()
+]
 
 
 def parse_ra(ra_str):
@@ -55,7 +62,7 @@ def build_constellations():
     constellation_records = []
     hiplines, hip_ids = read_hiplines()
 
-    for cid, props in constellations.properties.items():
+    for cid, props in properties.items():
         constellation_dict = {
             "iau_id": cid.lower(),
             "name": props[0].replace("\n", " "),
@@ -104,6 +111,10 @@ def build():
     constellation_records = build_constellations()
     df = pd.DataFrame.from_records(constellation_records)
 
+    for language, column_name in language_columns:
+        df[column_name] = df.apply(lambda d: translated[language].get(d.iau_id), axis=1)
+
+    print(df)
     gdf = gpd.GeoDataFrame(
         df,
         geometry=df["geometry"],