+++ examples/optic_solar_eclipse.py
@@ -22,14 +22,7 @@ op = m.create_optic(
     style=PlotStyle().extend(
         extensions.GRAYSCALE_DARK,
         extensions.OPTIC,
-        {
-            "background_color": [
-                (0.0, "#7abfff"),
-                (0.1, "#7abfff"),
-                (0.9, "#568feb"),
-                (0.9, "#3f7ee3"),
-            ]
-        },
+        extensions.GRADIENT_DAYLIGHT,
     ),
     resolution=2000,
 )