+++ src/starplot/models/__init__.py
@@ -1,5 +1,6 @@
 # ruff: noqa: F401,F403
 from .constellation import Constellation
+from .comet import Comet
 from .dso import DSO, DsoType
 from .star import Star
 from .planet import Planet
@@ -14,3 +15,5 @@ from .optics import (
     Camera,
 )
 from .objects import ObjectList
+from .satellite import Satellite
+from .observer import Observer