+++ src/starplot/optic.py
@@ -10,7 +10,7 @@ from starplot.base import BasePlot, DPI
 from starplot.data.stars import StarCatalog
 from starplot.mixins import ExtentMaskMixin
 from starplot.models import Star, Optic, Camera
-from starplot.observer import Observer
+from starplot.models.observer import Observer
 from starplot.plotters import (
     StarPlotterMixin,
     DsoPlotterMixin,
@@ -235,7 +235,7 @@ class OpticPlot(
         size_fn: Callable[[Star], float] = callables.size_by_magnitude_for_optic,
         alpha_fn: Callable[[Star], float] = callables.alpha_by_magnitude,
         color_fn: Callable[[Star], str] = None,
-        label_fn: Callable[[Star], str] = None,
+        label_fn: Callable[[Star], str] = Star.get_label,
         labels: Mapping[int, str] = None,
         legend_label: str = "Star",
         bayer_labels: bool = False,